+++ src/commands/make/Model.ts
@@ -19,19 +19,18 @@ export default class MakeModel extends Common {
         let isAll = false
 
         //Determine if should we create migration,factory and resource controller
-        isAll = await this.getYesNo('Should I create migration,factory and resource controller for the model?');
-        
-        if(!isAll)
-        {
+        isAll = await this.getYesNo('Should I create a migration, factory and resource controller for the model?');
+
+        if (!isAll) {
             // Determine if we should create a migration or not
             isMigration = await this.getYesNo('Should I create a migration for the model?');
 
-            // Determine if it should create a foctory for this model or not            
+            // Determine if it should create a factory for this model or not
             isFactory = await this.getYesNo('Should I create a factory for the model?');
-            
+
             // Should a controller be generated?
             isController = await this.getYesNo('Should I create a controller for the model?');
-    
+
             // Ask if the controller is a resource if the previous answer was 'yes'
             if (isController) {
                 // Determine if this is a resource controller or not
@@ -48,8 +47,15 @@ export default class MakeModel extends Common {
                 this.showError('Could not create the model', err);
             } else {
                 await this.openFile('/app/' + modelName + '.php');
-                if (isController) {
+                if (isController || isAll) {
                     await this.openFile('/app/Http/Controllers/' + modelName + 'Controller.php');
+                }
+                if (isFactory || isAll) {
+                    await this.openFile('/database/factories/' + modelName + 'Factory.php');
+                }
+                if (isMigration || isAll) {
+                    let migration = stdout.match(/Created Migration:(.+)/im)[1].trim()
+                    await this.openFile('/database/migrations/' + migration + '.php');
                 }
             }
         });